@@ -1,34 +1,31 @@
-import SECRETS from "./secrets.json";
-import type { TiltifyDonationProgress } from "./types";
+import SECRETS from './secrets.json';
+import type { TiltifyDonationProgress } from './types';
 
-export class Tiltify {
-    private static readonly _CAUSE_ID = "7886";
+const API_URL = 'https://tiltify.com/api/v3';
+const CAUSE_ID = '7886';
 
-    private static readonly _REQUIRED_HEADERS = new Headers({
-        "Content-Type": "text/json",
-        Authorization: `Bearer ${SECRETS.API_KEY}`,
-    });
+const REQUIRED_HEADERS = new Headers({
+  'Content-Type': 'text/json',
+  Authorization: `Bearer ${SECRETS.API_KEY}`,
+});
 
-    private static readonly _API_URL = "https://tiltify.com/api/v3";
-
-    public static getCurrentDonationProgress(): Promise<TiltifyDonationProgress> {
-        return fetch(
-            `${Tiltify._API_URL}/causes/${Tiltify._CAUSE_ID}/fundraising-events`,
-            {
-                method: "GET",
-                headers: Tiltify._REQUIRED_HEADERS,
-            }
-        ).then((response) => {
-            if (response.ok) {
-                return Promise.resolve(
-                    response.json().then((json) => ({
-                        current: json.data[0].totalAmountRaised,
-                        goal: json.data[0].fundraiserGoalAmount,
-                    }))
-                );
-            } else {
-                return Promise.reject("Response failed");
-            }
-        });
+export const getCurrentDonationProgress = async (): Promise<TiltifyDonationProgress> => {
+  return await fetch(
+    `${API_URL}/causes/${CAUSE_ID}/fundraising-events`,
+    {
+      method: 'GET',
+      headers: REQUIRED_HEADERS,
+    },
+  ).then(async (response) => {
+    if (response.ok) {
+      return await Promise.resolve(
+        response.json().then((json) => ({
+          current: json.data[0].totalAmountRaised,
+          goal: json.data[0].fundraiserGoalAmount,
+        })),
+      );
+    } else {
+      return await Promise.reject(new Error('Response failed'));
     }
-}
+  });
+};
