@@ -1,4 +1,4 @@
-export type TiltifyDonationProgress = {
-    current: number;
-    goal: number;
+export interface TiltifyDonationProgress {
+  current: number;
+  goal: number;
 }
